Measured power data in Qiling Line
The data derived from EMS record the power transmited in Qiling Line
which connects Daijiapu microgrid and the main grid. According to the
data, the curves of  Fig.2 could be drawn and the algorithm described in
the paper "Research on Configuration Strategy for Regional Energy
Storage System Based on Three Typical Filtering Methods" could also be
verified.